refactor: move filters to month select header
(cherry picked from commit 4a3d7ece803bbfaa84e5e12a40635833f1bfd8a8)
